mart_api/user_api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlmodel = "^0.0.22"
psycopg = {extras = ["binary"], version = "^3.2.3"}
psycopg2 = {extras = ["binary"], version = "^2.9.10"}
aiokafka = "^0.12.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
protobuf = "^5.28.3"
email-validator = "^2.2.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
